updated stream details page
